dflt_term_cap {fansi}	R Documentation

Default Arg Helper Funs

Description

Terminal capabilities are assumed to include bright and 256 color SGR codes. 24 bit color support is detected based on the COLORTERM environment variable.

Usage

dflt_term_cap()

dflt_css()

Details

Default CSS may exceed or fail to cover the interline distance when two lines have background colors. To ensure lines are exactly touching use inline-block, although that has its own issues. Otherwise specify your own CSS.

Value

character to use as default value for fansi parameter.

See Also

term_cap_test.

[Package fansi version 1.0.5 Index]