sgr_256 {fansi}	R Documentation

Show 8 Bit CSI SGR Colors

Description

Generates text with each 8 bit SGR code (e.g. the "###" in "38;5;###") with the background colored by itself, and the foreground in a contrasting color and interesting color (we sacrifice some contrast for interest as this is intended for demo rather than reference purposes).

Usage

sgr_256()

Value

character vector with SGR codes with background color set as themselves.

See Also

make_styles().

Examples

writeLines(sgr_256())

[Package fansi version 1.0.2 Index]